inventory_api/requirements.txt
asgiref==3.11.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
PyJWT==2.10.1
sqlparse==0.5.5
